numerify
removing _numerify

add mathics.core.evaluators module

CHANGES.rst
