+++ mathics/builtin/numbers/numbertheory.py
@@ -19,7 +19,7 @@ from mathics.core.atoms import (
 from mathics.core.convert import from_sympy, SympyPrime
 import mpmath
 
-from mathics.builtin.numeric import apply_N
+from mathics.core.evaluators import apply_N
 
 from mathics.core.attributes import (
     listable,
