+++ mathics/core/systemsymbols.py
@@ -25,7 +25,9 @@ SymbolInfix = Symbol("Infix")
 SymbolLess = Symbol("Less")
 SymbolMachinePrecision = Symbol("MachinePrecision")
 SymbolMakeBoxes = Symbol("MakeBoxes")
+SymbolMaxExtraPrecision = Symbol("$MaxExtraPrecision")
 SymbolMinus = Symbol("Minus")
+SymbolN = Symbol("N")
 SymbolNumberQ = Symbol("NumberQ")
 SymbolNumericQ = Symbol("NumericQ")
 SymbolPattern = Symbol("Pattern")
